test-ecdsa/requirements.txt
eth_keys==0.4.0
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.1 